requirements.txt
asgiref==3.8.1
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.1
django-ckeditor==6.7.1
django-crontab==0.7.1
django-environ==0.11.2
django-js-asset==2.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
idna==3.10
mysqlclient==2.2.4
openpyxl==3.1.5
pillow==10.4.0
psutil==6.0.0
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
